explicitly configure AWS SSM tests
this makes it easier to tell whether or not they should be run. it's
pretty likely for a developer to have ~/.aws configured on their machine
for some other reason, and not necessarily care about running these
tests or at least not want to globally reconfigure their machine for
them.

this also makes it easier to tell whether the session token needs to be
fetched in-suite or is already explicitly provided.